bizzer-website/backend/requirements/base.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Authentication
djoser>=2.2.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# Cache & Session
redis>=5.0.0
hiredis>=2.3.0

# Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Utils
python-decouple>=3.8
Pillow>=10.1.0
